Port OAM tiler code base to Raster Foundry
This is largely a copy and paste job, with a bunch of stray Bash script
and dummy data removals. The plan is to do an initial port that
successfully compiles into an assembly JAR and passes the test suite.
The second pass will actually attempt to get things working and fill out
the README in the process.